+++ src/main/resources/static/js/staff/stafflist.js
@@ -1,23 +1,7 @@
 $(document).ready(function(){
-    $("#search").on("keyup", function() {
-        let value = $(this).val().toLowerCase();
-        filter(value);
-    });
-
     $("#clear").on("click", function() {
        $("#search").val("").focus();
-       filter("");
+       return false;
     });
+    $("#search").select();
 });
-
-function filter(value) {
-    if (value.length == 0 || value.length >= 2) {
-        $("#staffList tbody tr").filter(function() {
-            let name = $(this).children().get(0).innerText.toLowerCase();
-            let legalName = $(this).children().get(1).innerText.toLowerCase();
-            let department = $(this).children().get(2).innerText.toLowerCase();
-            let visible = name.indexOf(value) > -1 || legalName.indexOf(value) > -1 || department.indexOf(value) > -1;
-            $(this).toggle(visible)
-        });
-    }
-}